Fix published types
